pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "model-based-software-engineering-assistant"
version = "0.1.0"
description = "A multi-agent LLM system for analyzing software models, generating code, and proposing refactorings"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["langchain", "langgraph", "multi-agent", "code-generation", "model-driven-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
requires-python = ">=3.10"

dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langgraph>=0.0.1",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=0.0.1",
    "faiss-cpu>=1.7.4",
    "chromadb>=0.4.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.10.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
    "pylint>=2.17.0"
]

jupyter = [
    "jupyter>=1.0.0",
    "ipython>=8.0.0",
    "ipywidgets>=8.0.0",
    "jupyterlab>=4.0.0"
]

api = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0"
]

observability = [
    "langsmith>=0.0.70"
]

all = [
    "model-based-software-engineering-assistant[dev,jupyter,api,observability]"
]

[project.urls]
Homepage = "https://github.com/yourusername/model-based-software-engineering-assistant"
Documentation = "https://github.com/yourusername/model-based-software-engineering-assistant#readme"
Repository = "https://github.com/yourusername/model-based-software-engineering-assistant"
Issues = "https://github.com/yourusername/model-based-software-engineering-assistant/issues"

[tool.setuptools]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | venv
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["backend"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests"
]

[tool.coverage.run]
source = ["backend"]
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@abstractmethod"
]
